runtime: add test for runtime drop issue
The test fails without the fix and passes with it.
It is run with rusty_fork in order to have a fresh process, so that we
can set a custom panic hook that aborts the process instead of
unwinding. This makes catching panics of tokio worker threads reliable.

Tokio worker threads panic if the runtime is dropped from an async
context, which is made more likely by the test's two properties:
1) all shared owners of the `Runtime` that are held by the main thread
   are dropped immediately, which makes more likely that the last owner
   is held by a tokio worker thread which executes the future.
2) the case is repeated 100 times in a loop, which increases the
   likelihood of the runtime being dropped from an async context.
   Without the loop, the test could sometimes pass even without the fix.