runtime: fix runtime drop panic
A CI run revealed a bug in the tokio runtime management: `Runtime`
might be dropped from an async context, which is forbidden (as it's
a blocking operation), so the worker thread panics.

Investigation showed that the problem manifests when a tokio task which
drives a future stored in `CassFuture` becomes the last shared owner
of the tokio `Runtime`, which is shared behind an `Arc`.

This commit fixes the problem by wrapping the `Runtime` within a custom
drop logic which switches `Runtime`'s' Drop behaviour from the blocking
shutdown to `shutdown_background()`. It intended to be used in places
where the runtime is closed from the context of another runtime.
Although our case is a bit different (we are closing the runtime from
the context of itself), I tested the approach and it seems to work well
and solve the problem.

Fixes: #362