testing: introduce rusty_fork_test_with_proxy!
The macro is similar to rusty_fork_test!, but it also passes a unique
IP address to the test function, which can be used to set up one proxy
node without risk of address collisions.

Without this macro, tests that use test_with_one_proxy() directly
and run in subprocesses would sometimes fail due to address collisions
with the main process or other subprocesses.