Cargo.toml: declare MSRV (1.85)
Since we already are using edition 2024, this workspace uses Cargo
resolver v3. Therefore, once we declare the MSRV, Cargo will
automatically ensure that all dependencies also support it.
Another benefit is that tools like `cargo clippy` will no longer
complain about possible improvements that are not supported by the MSRV.

The policy adopted is to follow MSRV of the Scylla Rust Driver, which is
currently set to 1.85.