docs: update MSRV requirements
As in the past we would take advantage of new Rust features, we prefer
to retain this flexibility by not declaring a specific MSRV in the
user-facing documentation. Instead, we now simply state that the
officially supported version is the most recent stable, and that older
versions may work, but we don't guarantee that.

This leaves us the freedom to use new Rust features as needed, by simply
bumping the MSRV in Cargo.toml as needed.