integration: proxy helper spawns blocking task
As the code in the closure passed to `test_with_3_node_dry_mode_cluster`
is allowed to perform blocking operations (and this often happens when
calling C API functions), we need to spawn a blocking task for it.

Now, the closure itself is not async anymore, as it is executed in a
blocking task. This avoids the pitfall of forgetting to spawn a
blocking task inside the closure, which would lead to panics in
certain situations.