integration: proxy helper accepts initial request rules
Although the only existing integration test - consistency.rs - does not
benefit from this change (as it must create request rules dynamically),
this change makes it easier to write future tests that do not need to
create request rules dynamically.
It will be used in subsequent commits, when we move more tests from
the lib crate to the integration tests.