testing_unimplemented: add missing `extern "C"`
The `extern "C"` block was missing, which is necessary for C linkage
of the functions in this file. The bug hadn't manifested yet because
these functions were declared in `cassandra.h` with `extern "C"`,
but it's best practice to have it in the implementation file as well.
This would manifest once we deleted the declarations in `cassandra.h`.