cassandra.h: add warnings for unimplemented functions
For each function in cassandra.h that is not yet implemented, add
a warning comment in its documentation to inform users.

We do not want to remove those functions from the header file,
as they are part of the public API and may be implemented in the future.

For now, use of them will result in linker errors, but at least no
runtime errors will occur.