Implement contact point shuffling
It is done in a simple way - by just shuffling the list before passing
it to Rust Driver. This matches the cpp-driver implementation.

A possibly better way is to implement ContronConnectionPolicy in Rust
Driver, and use it here to make shuffling possible for all hosts (not
only contact points) and re-shuffle each time CC is broken. This is a
greater effort, so for now lets go with cpp-driver impl.