codewide: migrate to unstable-cpp-rs feature
The recent change in the Rust driver replaced the `cpp_rust_unstable`
cfg with a more general `scylla_unstable` cfg to enable unstable
features, and added a new `unstable-cpp-rs` feature to enable
unstable APIs dedicated for CPP-RS Driver. This commit updates
the cpp-rs-driver to use the new cfg and feature accordingly.