implement cass_cluster_set_host_listener_callback
The host listener feature allows users to register a callback function
that will be invoked when certain events occur on the cluster's hosts,
such as when a host is added, removed, goes up, or goes down. This is
useful for monitoring the state of the cluster and reacting to changes
in real-time.

The host_listener example is unfortunately not suitable for automated
running in the CI, as it is designed to run indefinitely, listening
for host events.

I confirmed in manual tests that CPP-RS Driver's behaviour is in line
with CPP Driver's behaviour for host listener callbacks, by running
both drivers against a ccm-controlled ScyllaDB cluster and observing
that both drivers received the same host events when run host_listener.c
example.