test_metrics: Fix and enable MetricsTests.StatsConnections
Now that we support constant reconnect policy, there is no fundamental
problem with running the test. I had to change the numbers, because we
were always off by one. I'm mostly sure this is because Rust Driver
includes CC in metrics, but I did not investigate that in the code.